Autowiring of CoreViewMainContentMapper
