+++ src/lib/View/CoreViewMainContentMapper.php
@@ -6,36 +6,11 @@
 namespace EzSystems\HybridPlatformUi\View;
 
 use eZ\Publish\Core\MVC\Symfony\View\View;
-use EzSystems\HybridPlatformUi\Components\App;
-use EzSystems\HybridPlatformUi\Components\MainContent;
 use EzSystems\HybridPlatformUi\Mapper\MainContentMapper;
 
-class CoreViewMainContentMapper implements MainContentMapper
+/**
+ * Maps a CoreView to the App's MainContent.
+ */
+interface CoreViewMainContentMapper extends MainContentMapper
 {
-    /**
-     * @var \EzSystems\HybridPlatformUi\Components\App
-     */
-    private $app;
-
-    public function __construct(App $app)
-    {
-        $this->app = $app;
-    }
-
-    /**
-     * @param \eZ\Publish\Core\MVC\Symfony\View\View $view
-     */
-    public function map($view)
-    {
-        if (!$view instanceof View) {
-            throw new \InvalidArgumentException('Expected an \eZ\Publish\Core\MVC\Symfony\View\View');
-        }
-
-        $this->app->setConfig([
-            'mainContent' => [
-                'template' => $view->getTemplateIdentifier(),
-                'parameters' => $view->getParameters(),
-            ],
-        ]);
-    }
 }
